pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlp"
version = "0.1.0"
description = "Vector-based Label Propagation for low-rank graphs"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Anonymous"}]
license = {text = "MIT"}
dependencies = [
    "numpy",
    "scipy",
    "torch",
    "torch-geometric",
    "scikit-learn",
    "scikit-network",
    "tqdm",
    "pyyaml",
    "pandas",
    "psutil",
    "matplotlib"
]
